rework demo agents for clarity
