+++ agents/README.md
@@ -1,30 +1,38 @@
 # Agents
 
-Firstly enter into the Agents folder
-```
-cd examples.composabl.io\agents
-```
+In this folder you can find all the different example agents we support:
 
-Install poetry dependencies:
-```
+- **Airplane:** Learn how to implement an autopilot on an airplane
+- **Boiler:**
+- **CSTR:**
+- **Demo:**
+- **Filament Extruder:**
+- **Inventory:**
+- **Lunar Lander:**
+- **Maintenance:**
+- **Starship:**
+
+## Prerequisites
+
+- [Poetry](https://python-poetry.org/docs/)
+
+## Getting Started
+
+To make it easy, we are using [Poetry](https://python-poetry.org/docs/) as a dependency management tool that allows us to easily start up any of the agents.
+
+```bash
+# Install Dependencies
 poetry install
-```
 
-Then you can use any of these commands to execute an agent using poertry
-```
-poetry run cstr
-poetry run airplane
-poetry run starship
-poetry run boiler
-```
+# Set your License Key
+# if you do not have this, contact us at sales@composabl.io
+export COMPOSABL_KEY="YOUR_KEY"
 
-# Available Agents
+# Accept the EULA
+# note: by adding this variable, you are agreeing to our EULA at https://composabl.io/legal/sdk-beta-agreement
+export COMPOSABL_EULA_AGREED=1
 
-- main
-- starship 
-- filament_extruder 
-- cstr 
-- airplane 
-- inventory 
-- boiler 
-- maintenance
+# Run an agent
+# e.g., poetry run airplane, poetry run boiler, poetry run cstr, ...
+poetry run YOUR_AGENT
+```
