--- agents/demo/pyproject.toml
@@ -1,20 +1,0 @@
-[tool.poetry]
-name = "demo"
-version = "0.1.0"
-description = ""
-authors = ["Hunter Park <hunter@composabl.ai>"]
-readme = "README.md"
-packages = [{include = "demo"}]
-
-[tool.poetry.dependencies]
-python = "~3.8.13"
-composabl-dev = "*"
-
-torch = "2.0.0" # Latest version (2.0.1) has an issue with cuda installation (https://github.com/pytorch/pytorch/issues/100974)
-
-[tool.poetry.scripts]
-main = "demo.main:start"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
