--- agents/lunar_lander/pyproject.toml
@@ -1,19 +1,0 @@
-[tool.poetry]
-name = "agent-lunar-lander"
-version = "0.1.0"
-description = ""
-authors = ["Xavier Geerinck <xavier.geerinck@gmail.com>"]
-readme = "README.md"
-packages = [{include = "lunar_lander"}]
-
-[tool.poetry.dependencies]
-python = "~3.8"
-composabl = "^0.2.7"
-composabl-ray = "^0.2.7"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry.scripts]
-main = "lunar_lander.main:start"
