+++ agents/src/cstr/agent.py
@@ -1,0 +1,85 @@
+import os
+
+from composabl import Agent, Runtime, Scenario, Sensor, Skill
+
+from .teacher import CSTRTeacher
+
+license_key = os.environ["COMPOSABL_KEY"]
+
+
+def start():
+    print("composabl_core|====================================================================")
+    print("composabl_core|")
+    print("composabl_core|Running the CSTR Agent")
+
+    T = Sensor("T", "")
+    Tc = Sensor("Tc", "")
+    Ca = Sensor("Ca", "")
+    Cref = Sensor("Cref", "")
+    Tref = Sensor("Tref", "")
+
+    sensors = [T, Tc, Ca, Cref, Tref]
+
+    # Cref_signal is a configuration variable for Concentration and Temperature setpoints
+    ss1_scenarios = [
+        {
+            "Cref_signal": "ss1"
+        }
+    ]
+
+    ss2_scenarios = [
+        {
+            "Cref_signal": "ss2"
+        }
+    ]
+
+    transition_scenarios = [
+        {
+            "Cref_signal": "transition"
+        }
+    ]
+
+    selector_scenarios = [
+        {
+            "Cref_signal": "complete"
+        }
+    ]
+
+    ss1_skill = Skill("ss1", CSTRTeacher, trainable=True)
+    for scenario_dict in ss1_scenarios:
+        ss1_skill.add_scenario(Scenario(scenario_dict))
+
+    ss2_skill = Skill("ss2", CSTRTeacher, trainable=True)
+    for scenario_dict in ss2_scenarios:
+        ss2_skill.add_scenario(Scenario(scenario_dict))
+
+    transition_skill = Skill("transition", CSTRTeacher, trainable=True)
+    for scenario_dict in transition_scenarios:
+        transition_skill.add_scenario(Scenario(scenario_dict))
+
+    selector_skill = Skill("selector", CSTRTeacher, trainable=True)
+    for scenario_dict in selector_scenarios:
+        selector_skill.add_scenario(Scenario(scenario_dict))
+
+    config = {
+        "env": {
+            "name": "sim-cstr",
+            "compute": "local",  # "docker", "kubernetes", "local"
+            "config": {
+                "address": "localhost:1337",
+                # "image": "composabl/sim-cstr:latest"
+            }
+        },
+        "license": license_key,
+        "training": {}
+    }
+    runtime = Runtime(config)
+    agent = Agent(runtime, config)
+    agent.add_sensors(sensors)
+
+    agent.add_skill(ss1_skill)
+    agent.add_skill(ss2_skill)
+    agent.add_skill(transition_skill)
+    agent.add_selector_skill(selector_skill, [ss1_skill, transition_skill, ss2_skill], fixed_order=False, fixed_order_repeat=False)
+
+    agent.train(train_iters=3)
