+++ agents/src/demo/teacher.py
@@ -1,0 +1,38 @@
+from composabl import Teacher
+
+
+class ReachTeacher(Teacher):
+    def __init__(self):
+        self.obs_history = None
+        self.reward_history = []
+        self.last_reward = 0
+
+    def transform_obs(self, obs, action):
+        return obs
+
+    def transform_action(self, transformed_obs, action):
+        return action
+
+    def filtered_observation_space(self):
+        return ["state1", "time_counter"]
+
+    def compute_reward(self, transformed_obs, action):
+        print(transformed_obs)
+        if self.obs_history is None:
+            self.obs_history = [transformed_obs]
+            return 0
+        else:
+            self.obs_history.append(transformed_obs)
+
+        reward = transformed_obs[0]
+        return reward
+
+    def compute_action_mask(self, transformed_obs, action):
+        return None
+
+    def compute_success_criteria(self, transformed_obs, action):
+        return len(self.obs_history) > 1000
+
+    def compute_termination(self, transformed_obs, action):
+        return False
+
