--- agents/train_agent/inventory_teacher.py
@@ -1,45 +1,0 @@
-import math
-
-from composabl_core.agent import Teacher
-
-class BalanceTeacher(Teacher):
-    def __init__(self):
-        self.obs_history = None
-        self.reward_history = []
-        self.last_reward = 0
-
-    def transform_obs(self, obs, action):
-        return obs
-
-    def transform_action(self, transformed_obs, action):
-        return action
-
-    def filtered_observation_space(self):
-        return ["inventory",
-                "balance",
-                "num_ordered",
-                "holding_cost",
-                "cost_price",
-                "delay_days_until_delivery",
-                "customer_demand_min",
-                "customer_demand_max",
-                "selling_price"]
-
-    def compute_reward(self, transformed_obs, action):
-        if self.obs_history is None:
-            self.obs_history = [transformed_obs]
-            return 0
-        else:
-            self.obs_history.append(transformed_obs)
-
-        reward = transformed_obs["balance"]
-        return reward
-
-    def compute_action_mask(self, transformed_obs, action):
-        return None
-
-    def compute_success_criteria(self, transformed_obs, action):
-        return len(self.obs_history) > 100
-
-    def compute_termination(self, transformed_obs, action):
-        return False
