--- agents/train_agent/perceptors.py
@@ -1,1 +1,0 @@
-from composabl.agent import Perceptor
